Reverted expected/actual ordering Moved common set up/tear down to test.test_common_set_up
